Add rotate method to `BluemiraGeo` and circular_pattern to tools (#444)
* add rotate and tests

* add circular patterning

* cleanup and centroid test

* docstrings

+++ bluemira/geometry/_freecadapi.py
@@ -654,16 +654,6 @@ def save_as_STEP(shapes, filename="test", scale=1):
         # a copy of the compound is made to avoid modification of the original shapes.
         compound = compound.copy().scale(scale)
 
-    # doc = FreeCAD.newDocument()
-    # obj = FreeCAD.ActiveDocument.addObject("App::DocumentObject", "Test")
-    #
-    # freecad_comp = FreeCAD.ActiveDocument.addObject("Part::Feature")
-    #
-    # # link the solid to the object
-    # freecad_comp.Shape = compound
-    #
-    # Part.export([freecad_comp], filename)
-
     compound.exportStep(filename)
 
 
@@ -723,7 +713,7 @@ def rotate_shape(
         Origin location of the rotation
     direction: tuple (x,y,z)
         The direction vector
-    degree: double
+    degree: float
         rotation angle
 
     Returns
